requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
chardet==3.0.4
Click==7.0
discord.py==1.0.1
dnspython==1.15.0
dnspython3==1.15.0
idna==2.8
idna-ssl==1.1.0
mcstatus==2.2
more-itertools==7.0.0
multidict==4.5.2
pluggy==0.9.0
py==1.8.0
pytest==4.4.1
python-dotenv==0.10.1
six==1.12.0
typing-extensions==3.7.2
websockets==6.0
yarl==1.3.0
